refactor: add `payment_id`, `validated`, `note`
payment_id: is related to payments table
validated: is the final result of exam, doctor decide to mark the exam result is valid or invalid?
note: note for patient, why their exam result is valid or invalid and doctor should be given a note how to get better score. Note is written by using wysiwyg editor